This commit segregate the functions into corresponding packages
