+++ pkg/apis/openebs.io/v1alpha1/cas_keys.go
@@ -96,6 +96,9 @@ const (
 
 	//PodDisruptionBudgetKey is the key used to identify the PDB
 	PodDisruptionBudgetKey = "openebs.io/pod-disruption-budget"
+
+	// CstorpoolInstanceLabel is the key used on pool dependent resources
+	CstorpoolInstanceLabel = "cstorpoolinstance.openebs.io/name"
 )
 
 // CASPlainKey represents a openebs key used either in resource annotation
