+++ pkg/kubernetes/poddisruptionbudget/poddisruptionbudget.go
@@ -1,0 +1,26 @@
+/*
+Copyright 2019 The OpenEBS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+// GetPDBLabelSelector returns the labelSelector to list the PDB
+func GetPDBLabelSelector(pdbLabels map[string]string) string {
+	var labelSelector string
+	for key, value := range pdbLabels {
+		labelSelector = labelSelector + key + "=" + value + ","
+	}
+	return labelSelector[:len(labelSelector)-1]
+}
